rename that files containing compressors starts with "compressor_<type>[.h/.cc]"
