+++ cvmfs/CMakeLists.txt
@@ -58,11 +58,11 @@ if (BUILD_CVMFS OR BUILD_LIBCVMFS)
        catalog_sql.cc
        clientctx.cc
        compression/compression.cc
-       compression/echo.cc
+       compression/compressor_echo.cc
+       compression/compressor_zlib.cc
        compression/input_file.cc
        compression/input_mem.cc
        compression/input_path.cc
-       compression/zlib.cc
        directory_entry.cc
        duplex_fuse.cc
        fd_refcount_mgr.cc
@@ -142,11 +142,11 @@ if (BUILD_CVMFS)
   #
   add_executable (cvmfs_fsck
                   compression/compression.cc
-                  compression/echo.cc
+                  compression/compressor_echo.cc
+                  compression/compressor_zlib.cc
                   compression/input_file.cc
                   compression/input_mem.cc
                   compression/input_path.cc
-                  compression/zlib.cc
                   cvmfs_fsck.cc
                   statistics.cc
   )
@@ -506,11 +506,11 @@ if (BUILD_LIBCVMFS_CACHE)
                   cache.cc
                   cache_posix.cc
                   compression/compression.cc
-                  compression/echo.cc
+                  compression/compressor_echo.cc
+                  compression/compressor_zlib.cc
                   compression/input_file.cc
                   compression/input_mem.cc
                   compression/input_path.cc
-                  compression/zlib.cc
                   fd_refcount_mgr.cc
                   manifest.cc
                   quota.cc
@@ -589,11 +589,11 @@ if (BUILD_SERVER)
        catalog_virtual.cc
        clientctx.cc
        compression/compression.cc
-       compression/echo.cc
+       compression/compressor_echo.cc
+       compression/compressor_zlib.cc
        compression/input_file.cc
        compression/input_mem.cc
        compression/input_path.cc
-       compression/zlib.cc
        directory_entry.cc
        file_chunk.cc
        gateway_util.cc
@@ -738,11 +738,11 @@ if (BUILD_SERVER)
        catalog_mgr_rw.cc
        catalog_virtual.cc
        compression/compression.cc
-       compression/echo.cc
+       compression/compressor_echo.cc
+       compression/compressor_zlib.cc
        compression/input_file.cc
        compression/input_mem.cc
        compression/input_path.cc
-       compression/zlib.cc
        directory_entry.cc
        gateway_util.cc
        globals.cc
@@ -853,11 +853,11 @@ if (BUILD_SERVER)
   set (CVMFS_PUBLISH_SOURCES
        backoff.cc
        compression/compression.cc
-       compression/echo.cc
+       compression/compressor_echo.cc
+       compression/compressor_zlib.cc
        compression/input_file.cc
        compression/input_mem.cc
        compression/input_path.cc
-       compression/zlib.cc
        directory_entry.cc
        manifest.cc
        network/dns.cc
@@ -1040,11 +1040,11 @@ if(BUILD_RECEIVER)
        catalog_mgr_ro.cc
        catalog_mgr_rw.cc
        compression/compression.cc
-       compression/echo.cc
+       compression/compressor_echo.cc
+       compression/compressor_zlib.cc
        compression/input_file.cc
        compression/input_mem.cc
        compression/input_path.cc
-       compression/zlib.cc
        directory_entry.cc
        gateway_util.cc
        globals.cc
@@ -1207,11 +1207,11 @@ if (BUILD_PRELOADER)
                   catalog.cc
                   catalog_sql.cc
                   compression/compression.cc
-                  compression/echo.cc
+                  compression/compressor_echo.cc
+                  compression/compressor_zlib.cc
                   compression/input_file.cc
                   compression/input_mem.cc
                   compression/input_path.cc
-                  compression/zlib.cc
                   gateway_util.cc
                   globals.cc
                   history_sql.cc
