+++ cvmfs/compression/compressor_echo.h
