+++ cvmfs/compression/compressor_zlib.cc
@@ -1,15 +1,10 @@
 /**
  * This file is part of the CernVM File System.
  *
- * This is a wrapper around zlib.  It provides
- * a set of functions to conveniently compress and decompress stuff.
- * Almost all of the functions return true on success, otherwise false.
- *
- * TODO: think about code deduplication
  */
 
 
-#include "zlib.h"
+#include "compressor_zlib.h"
 
 #include <alloca.h>
 #include <stdlib.h>
