+++ test/micro-benchmarks/CMakeLists.txt
@@ -19,11 +19,11 @@ set (CVMFS_UBENCHMARKS_SOURCES
   # dependencies
   ${CVMFS_SOURCE_DIR}/cache_transport.cc
   ${CVMFS_SOURCE_DIR}/compression/compression.cc
-  ${CVMFS_SOURCE_DIR}/compression/echo.cc
+  ${CVMFS_SOURCE_DIR}/compression/compressor_echo.cc
+  ${CVMFS_SOURCE_DIR}/compression/compressor_zlib.cc
   ${CVMFS_SOURCE_DIR}/compression/input_file.cc
   ${CVMFS_SOURCE_DIR}/compression/input_mem.cc
   ${CVMFS_SOURCE_DIR}/compression/input_path.cc
-  ${CVMFS_SOURCE_DIR}/compression/zlib.cc
   ${CVMFS_SOURCE_DIR}/crypto/hash.cc
   ${CVMFS_SOURCE_DIR}/directory_entry.cc
   ${CVMFS_SOURCE_DIR}/glue_buffer.cc
