+++ test/stress/CMakeLists.txt
@@ -3,11 +3,11 @@ set (CVMFS_SOURCE_DIR "${CMAKE_SOURCE_DIR}/cvmfs")
 add_executable(s3benchmark
                test/stress/s3benchmark.cc
                ${CVMFS_SOURCE_DIR}/compression/compression.cc
-               ${CVMFS_SOURCE_DIR}/compression/echo.cc
+               ${CVMFS_SOURCE_DIR}/compression/compressor_echo.cc
+               ${CVMFS_SOURCE_DIR}/compression/compressor_zlib.cc
                ${CVMFS_SOURCE_DIR}/compression/input_file.cc
                ${CVMFS_SOURCE_DIR}/compression/input_mem.cc
                ${CVMFS_SOURCE_DIR}/compression/input_path.cc
-               ${CVMFS_SOURCE_DIR}/compression/zlib.cc
                ${CVMFS_SOURCE_DIR}/gateway_util.cc
                ${CVMFS_SOURCE_DIR}/json_document.cc
                ${CVMFS_SOURCE_DIR}/network/dns.cc
