+++ test/unittests/CMakeLists.txt
@@ -39,11 +39,11 @@ if (BUILD_LIBCVMFS_CACHE)
                   ${CVMFS_SOURCE_DIR}/cache_extern.cc
                   ${CVMFS_SOURCE_DIR}/cache_transport.cc
                   ${CVMFS_SOURCE_DIR}/compression/compression.cc
-                  ${CVMFS_SOURCE_DIR}/compression/echo.cc
+                  ${CVMFS_SOURCE_DIR}/compression/compressor_echo.cc
+                  ${CVMFS_SOURCE_DIR}/compression/compressor_zlib.cc
                   ${CVMFS_SOURCE_DIR}/compression/input_file.cc
                   ${CVMFS_SOURCE_DIR}/compression/input_mem.cc
                   ${CVMFS_SOURCE_DIR}/compression/input_path.cc
-                  ${CVMFS_SOURCE_DIR}/compression/zlib.cc
                   ${CVMFS_SOURCE_DIR}/manifest.cc
                   ${CVMFS_SOURCE_DIR}/quota.cc
   )
@@ -163,11 +163,11 @@ if (BUILD_SERVER)
        ${CVMFS_SOURCE_DIR}/catalog_mgr_rw.cc
        ${CVMFS_SOURCE_DIR}/catalog_virtual.cc
        ${CVMFS_SOURCE_DIR}/compression/compression.cc
-       ${CVMFS_SOURCE_DIR}/compression/echo.cc
+       ${CVMFS_SOURCE_DIR}/compression/compressor_echo.cc
+       ${CVMFS_SOURCE_DIR}/compression/compressor_zlib.cc
        ${CVMFS_SOURCE_DIR}/compression/input_file.cc
        ${CVMFS_SOURCE_DIR}/compression/input_mem.cc
        ${CVMFS_SOURCE_DIR}/compression/input_path.cc
-       ${CVMFS_SOURCE_DIR}/compression/zlib.cc
        ${CVMFS_SOURCE_DIR}/directory_entry.cc
        ${CVMFS_SOURCE_DIR}/gateway_util.cc
        ${CVMFS_SOURCE_DIR}/globals.cc
@@ -432,11 +432,11 @@ set(CVMFS_UNITTEST_SOURCES
   ${CVMFS_SOURCE_DIR}/catalog_virtual.cc
   ${CVMFS_SOURCE_DIR}/clientctx.cc
   ${CVMFS_SOURCE_DIR}/compression/compression.cc
-  ${CVMFS_SOURCE_DIR}/compression/echo.cc
+  ${CVMFS_SOURCE_DIR}/compression/compressor_echo.cc
+  ${CVMFS_SOURCE_DIR}/compression/compressor_zlib.cc
   ${CVMFS_SOURCE_DIR}/compression/input_file.cc
   ${CVMFS_SOURCE_DIR}/compression/input_mem.cc
   ${CVMFS_SOURCE_DIR}/compression/input_path.cc
-  ${CVMFS_SOURCE_DIR}/compression/zlib.cc
   ${CVMFS_SOURCE_DIR}/cvmfs_suid_util.cc
   ${CVMFS_SOURCE_DIR}/directory_entry.cc
   ${CVMFS_SOURCE_DIR}/duplex_fuse.cc
